Use bigint instead of integer.
Integer is too small for the 64 bit types used by postgres.

+++ src/PostgresWriter.cc
@@ -57,7 +57,7 @@ string PostgreSQL::GetTableType(int arg_type, int arg_subtype)
 	case TYPE_COUNT:
 	case TYPE_COUNTER:
 	case TYPE_PORT:
-		type = "integer";
+		type = "bigint";
 		break;
 
 	/*
